Unfull game
This adds a button to the drop-down box on the left that will let you change the number of players in the game. This is just to save people time so they don't have to mark someone as dead at the start of the game over and over, I will be enhancing this feature for a little while as I am not entirely happy with it.